+++ packages/docs/site/src/routes/compounds/time_region_value/index.svelte
@@ -1,8 +1,6 @@
 <script>
 	import Index from '@svizzle/time_region_value/src/routes/index.svelte';
 
-	import {goto as goTo} from '@sapper/app';
-
 	import {_groups, _yearRange} from './_data/stores';
 	import {hrefBase} from './_navigation';
 </script>
@@ -10,6 +8,5 @@
 <Index
 	{_groups}
 	{_yearRange}
-	{goTo}
 	{hrefBase}
 />